+++ evenement.php
@@ -51,8 +51,7 @@ if (!$even->getValues() || in_array($even->getValue('statut'), ['inactif', 'prop
 	else
 	{
 		header("HTTP/1.1 404 Not Found");
-		echo file_get_contents("articles/404.php");
-		exit;
+        exit;
 	}
 }