+++ lieu.php
@@ -24,8 +24,7 @@ if (isset($_GET['idL']) && (int) $_GET['idL'] > 0)
 else
 {
 	header("HTTP/1.1 404 Not Found");
-    echo file_get_contents("articles/404.php");
-	exit;
+    exit;
 }
 
 $tab_genre_even = ["fête", "cinéma", "théâtre", "expos", "divers", "tous"];
@@ -74,7 +73,6 @@ $lieu->load();
 if ($lieu->getValue('statut') == 'inactif' && !((isset($_SESSION['Sgroupe']) && $_SESSION['Sgroupe'] <= 6)))
 {
     header("HTTP/1.1 404 Not Found");
-    echo file_get_contents("articles/404.php");
     exit;
 }
 
@@ -131,7 +129,7 @@ elseif ($lieu->getValue('statut') == 'inactif')
 	else
 	{
         header("HTTP/1.1 404 Not Found");
-        echo file_get_contents("articles/404.php");
+        echo file_get_contents("articles/error.php");
         exit;
 	}
 }